chatbot-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.4
grpcio==1.70.0
grpcio-tools==1.70.0
transformers==4.38.2
torch==2.2.2
protobuf==5.29.0
